+++ metis-os-workspace/metis-settings/src/pages/mod.rs
@@ -1,5 +1,5 @@
 pub mod appearance;
 pub mod calendars;
+pub mod menu;
 pub mod network;
-pub mod startmenu;
 pub mod weather;